chatgpt_project_eval/requirements.txt
pyautogui==0.9.54
openai>=1.30.0
pillow>=10.0.0
